ENH: Interactive DensityProfile PV plugin
Added a paraview filter that allows the user to interactively
inspect a particle dataset and plot a density profile (i.e.,
the number of particles within a region). The user picks a
center, a sphere radius and the number of concentric spheres
to generate. For each sphere, the number of particles is
computed. When running in parallel, each process computes the
number of particles within each sphere independently and the
final results are sum-reduced to process 0 which is responsible
for generating the final table. In the present implementation the
overlap should be set to 0, i.e. no ghost cells to guarrantee
correctness.