ENH: Density Profile of FOF Centers (Part 1)
Initial changes to the density profile to allow calculation
of the profile at FOF centers. The user controls whether or
not to interactively pick the center or to use the FOF centers
for computing a density profile via a checkbox in the filter's
properties panel. Internally, the code was modified s.t. it
operates on an array of SphereCenters. When interactive there
is only a single SphereCenter which the user has access to control
and move about the domain. Otherwise, when UseFOFCenters is ON,
the SphereCenters array is populated with FOF centers returned
from the FOF halo finder.