ENH: Density Profile of FOF Centers (Part 2)
Changed filter such that it has two input ports for its two possible
inputs. The first input is required and consists of the raw particles.
The second input is optional and consists of the pre-calculated FOF
centers. If the FOF centers are provided, then, the filter will generate
a profile for each FOF. Otherwise, the user has the ability to interactively
move the cursor to a desired center.