COMP: Propagate package include dirs and libs
Changed build such that each package stores its include dirs
and libraries in a variable PACKAGE_${pkg}_INCLUDE_DIRS and
PACKAGE_${pkg}_LIBRARIES respectively. These variables are
INTERNAL CACHE variables so other projects can use them, set
them etc. To achieve this the build had to be changed to be
less dynamic to resolve difficulties with dependencies etc.
At some point we will go back and fix this, but, for now,
dependencies are explicitely set by the order projects are
added in the main CMakeLists.txt.