ENH: Refactoring of LANL halo-finder (Part 1)
This commit includes changes to the LANL halo-finder that:

(1) Get's rid of the 800++ line-of-code RequestData method that was
hard to maintain and modify. Instead, the code is logicaly decomposed
into modular parts

(2) Logic is added to keep an internal mapping of global particle IDs
to local particle IDs in this process.

(3) Removed the ComputeMBP and ComputeMCP flags. Instead, a CenterFindingMethod
ivar is used to allow the user to choose from the 4 center-finding alogorithms
that are available.

(4) Corrected how the halo-centers were constructed. Previous implementation was
always storing the average particle position in the halo-center regardless of
what center-finding method was used. Now, the halo-centers will reflect the actual
center-finding algorithm that is employed.

Note, this commit has essentially re-implemented the filter. SOD-halos are currently
not available.