ENH: Euler mesh probing implementation
Added functionality to probe the euler mesh with a query point P and
find the number of streams at the given point and compute an estimate
of the local density rho. After the euler mesh is constructed, the
tetrahedra are injected to a virtual grid, i.e., a chaining mesh to
speed-up these queries.