ENH: Use internal ivars to store filter outputs
Use internal ivars to store the filter's output which are shallow copied
each time the RequestData is called. This is done b/c each time RequestData
is called, not all outputs of the filter may be re-generated. For example, if
the probe grid parameters change, only the probe grid is updated -- the previously
generated euler mesh and caustic surface is used.