ENH: Base co-processing pipeline implementation
Implemented co-processing pipeline. When CoProcess is called,
the configuration file is parsed and concrete instances of
analysis tools are created. The main code then loops through
all analysis tools and for each tool it performs three actions:
(1) Calls Execute to run the underlying analysis algorithm
(2) If generate output was requested it will call the analysis
algorithm WriteOutput routine
(3) If visualization is ON, it will send the data to the visualization
server.

Currently, support only for the LANL halo-finder is implemented. Further,
on-the-fly visualization is not yet implemented.