DOC: Add sample config for halo-tracker
The halo-tracker takes essentially the same parameters
as the halo-finder + 2 more parameters to prescribe
the merger-tree threshold used to determine which
halos are the same and the merger-tree file that
is used to dump the DistributedHaloEvolutionTree
instance at the end of the simulation. Also the
INSTANCE_NAME changes from 'LANLHALOFINDER' to
'HALOTRACKER' to indicate to the class instantiator
which class to instantiate.