ENH:Implement MPIUtilities
Implement MPIUtilities singleton class that provides functionality
for printing to stdout, either on the master node, i.e., rank 0, or
in a synchronized manner, that is, rank i, will print before rank i+1.
This class also serves as the place holder for other similar functions
to be added in the future.