COMP: Define HDR_ELEMS if not defined
Latest changes in tess do not define a HDR_ELEMS, however, when this
reader was initially written, HDR_ELEMS was explicitly defined and
was being used internally. Consequently, the reader was not being
compiled properly. To rectify the issue, a check was added to check
if the HDR_ELEMS is defined and if it isn't the code would explicitely
define it. In the future, this reader needs to be updated however to
use the DIY I/O routines (which have been updated) to read the file.