Third attempt, this time we set virtual grid to be same as serial ver.
Setting the virtual grid to be based on the euler mesh instead
of the probe grid we can improve the subdivision since more of
virtual grids cells will have tets inside them.

This also makes the code far closer in nature to the serial implementation
which makes it a fair comparison.